[issue/39] fix build-queue toggle test
